feat: add image handling middleware, Swagger setup, and AdminJS integration; update server configuration
