+++ backend/src/server.ts
@@ -1,8 +1,38 @@
-// src/server.js
-
 import app from './app.js';
+import { connect } from 'mongoose';
 import logger from './services/logger.js';
 import { socketService } from './services/socket.js';
+import swaggerSetup from './modules/swagger.js';
+import adminJsSetup from './modules/admin.js';
+import { exit } from 'process';
+import { isProduction, rootDir } from './config.js';
+import { createWriteStream } from 'fs';
+import { join } from 'path';
+import morgan from 'morgan';
+
+// Database Connection
+if (process.env.MONGO_URI == null) {
+    logger.error('MONGO_URI is not defined');
+    exit(1);
+}
+
+// Logging
+if (isProduction) {
+    console.log('lancement en production');
+    const accessLogStream = createWriteStream(join(rootDir, 'logs', 'access.log'), { flags: 'a+' });
+    app.use(morgan('combined', { stream: accessLogStream }));
+} else {
+    console.log('lancement en dev');
+    app.use(morgan('dev'));
+}
+
+logger.info('MONGO_URI: ' + process.env.MONGO_URI);
+connect(process.env.MONGO_URI)
+    .then(() => logger.info('MongoDB Connected'))
+    .catch(err => logger.error('MongoDB connection error:', err));
+
+swaggerSetup(app);
+adminJsSetup(app);
 
 const PORT = process.env.PORT || 5000;
