--- backend/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-    test: {
-        globals: true,
-        environment: 'node',
-    },
-});
